Switch to git using git/bzr connector
New patch to allow debugging with pydev configurable through environment variables

Note: in my environment the initial clone for schooltool project fails if executed inside the container. So, I needed to run "git clone bzr::lp:~schooltool-owners/schooltool/ark schooltool" in the host (in sandbox directory). For this, the git/bzr connecter is needed in the host (apt-get install git-remote-bzr).